+++ dataprovider/baseballreferencemlb/scraper_matchups.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/lightning-dabbler/sportscrape"
 	"github.com/lightning-dabbler/sportscrape/dataprovider/baseballreferencemlb/model"
+	"github.com/lightning-dabbler/sportscrape/scraper"
 	"github.com/lightning-dabbler/sportscrape/util"
 	"github.com/lightning-dabbler/sportscrape/util/sportsreference"
 	"github.com/xitongsys/parquet-go/types"
@@ -67,7 +68,7 @@ func NewMatchupScraper(options ...MatchupOption) *MatchupScraper {
 
 // MatchupScraper specialized Runner for retrieving MLB matchup information.
 type MatchupScraper struct {
-	sportsreference.BaseScraper
+	scraper.BaseScraper
 	Date string
 }